Moved Backend to Folder for Integration Cohesion
