+++ api/frontMockup/index.html
@@ -1,0 +1,57 @@
+<html>
+
+<head>
+    <title>UF Parking Map Plus</title>
+    <link rel="icon" type="image/x-icon" href="favicon/icons8-map-marker-material-filled-32.ico">
+</head>
+
+<body>
+    <div style="display: flex">
+        <h1 style="display: flex; flex: 0.5; margin-left: 20px">UF Parking Map Plus</h1>
+        <form action="/map/search" method="POST" style="display:flex; flex: 1;">
+            <label for="/map/search" style="display:flex; align-items: center; padding-right: 5px;">Address or Building Name:</label>
+            <input id=/map/search name="/map/search" type="search" placeholder="Address or Building" style="width: 50%;">
+        </form>
+    </div>
+
+    <div>
+        <iframe src="https://www.google.com/maps/embed?pb=!1m14!1m12!1m3!1d9807.758111021383!2d-82.35676600118562!3d29.645344113047244!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!5e0!3m2!1sen!2sus!4v1675485248160!5m2!1sen!2sus"
+                width="100%" height="450" style="border:1;" allowfullscreen="" loading="lazy"
+                referrerpolicy="no-referrer-when-downgrade"></iframe>
+    </div>
+
+    <div>
+        <h2 style="text-align: center">Filters</h2>
+
+        <form style="text-align: center" action="/filter/pd" method="POST">
+            <label for="filter/pd">Parking Decal:</label>
+            <select name="filter/pd", id="filter/pd" onchange="this.form.submit()">
+                <option value="green"> Green </option>
+                <option value="parkride"> Park & Ride</option>
+                <option value="red1"> Red 1</option>
+                <option value="red3"> Red 3</option>
+                <option value="brown2"> Brown 2</option>
+                <option value="brown3"> Brown 3</option>
+                <option value="disabledStudent"> Disabled Student </option>
+                <option value="motorcycleScooter"> Motorcycle/Scooter </option>
+                <option value="gold"> Gold </option>
+                <option value="silver"> Silver </option>
+                <option value="officialBusiness"> Official Business</option>
+                <option value="orange"> Orange </option>
+                <option value="blue"> Blue </option>
+                <option value="medicalResident"> Medical Resident </option>
+                <option value="staffCommuter"> Staff Commuter </option>
+                <option value="disabledEmployee"> Disabled Employee </option>
+                <option value="carpool"> Carpool </option>
+                <option value="commercial"> Commercial </option>
+            </select>
+        </form>
+    </div>
+</body>
+<footer style="text-align: center">
+    DISCLAIMER: This web application is merely a tool. Always follow all official signage.
+    <br>
+    <a target="_blank" href="https://icons8.com/icon/85879/map-marker">Map Marker</a> icon by <a target="_blank" href="https://icons8.com">Icons8</a>
+</footer>
+
+</html>
