--- go.mod
@@ -1,11 +1,0 @@
-module ufpmp
-
-go 1.19
-
-require (
-	github.com/felixge/httpsnoop v1.0.3 // indirect
-	github.com/gorilla/handlers v1.5.1 // indirect
-	github.com/gorilla/mux v1.8.0 // indirect
-	github.com/pkg/browser v0.0.0-20210911075715-681adbf594b8 // indirect
-	golang.org/x/sys v0.0.0-20210616045830-e2b7044e8c71 // indirect
-)
