--- httpd/app_handler/app_handler.go
@@ -1,14 +1,0 @@
-package app_handler
-
-import (
-	"fmt"
-	"log"
-	"net/http"
-)
-
-func handler(w http.ResponseWriter, r *http.Request) {
-	_, err := fmt.Fprint(w, "Main Page")
-	if err != nil {
-		log.Fatal(err)
-	}
-}
